+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.90.0] - 2021-09-25
+* Added cmdlet to generate mermaid diagram to be used in MarkDown type of documents
+* Added public command `Start-AdfTriggers`
+* Added retry action on failure of starting trigger #107
+
 ## [0.80.0] - 2021-08-17
 * Fixed `Get-ReferencedObjects` which fails for files with empty element in JSON #128